# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for email verification service",
    "title": "Email Checker API",
    "version": "1.0",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    },
    "AdminKeyAuth": {
      "type": "apiKey",
      "name": "X-Admin-Key",
      "in": "header"
    }
  },
  "paths": {

    "/admin/keys": {
      "get": {
        "summary": "List all API keys",
        "tags": ["Administration"],
        "responses": {
          "200": {
            "description": "List of API keys",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FullAPIKeyDetails"
              }
            }
          }
        }
      }
    },"/admin/keys/{api_key}": {
      "get": {
        "summary": "Get key details",
        "parameters": [
          {
            "name": "api_key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Key details",
            "schema": {
              "$ref": "#/definitions/FullAPIKeyDetails"
            }
          }
        }
      },
      "patch": {
        "summary": "Update key",
        "parameters": [
          {
            "name": "api_key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/KeyUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update status",
            "schema": {
              "type": "object",
              "properties": {
                "status": {"type": "string"}
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete key",
        "responses": {
          "204": {
            "description": "Key deleted"
          }
        }
      }
    },
    "/cache/status": {
      "get": {
        "summary": "Get cache status",
        "description": "Get cache statistics",
        "tags": ["cache"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Cache statistics",
            "schema": {
              "$ref": "#/definitions/CacheStatusResponse"
            }
          }
        }
      }
    },
    "/cache/flush": {
      "post": {
        "summary": "Clear cache",
        "description": "Flush all cached data",
        "tags": ["cache"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Cache successfully flushed"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "post": {
        "summary": "Create new verification task",
        "description": "Create task for email verification",
        "tags": ["tasks"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Email list",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TaskIDResponse"
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "summary": "Get task status",
        "description": "Get verification task status and results",
        "tags": ["tasks"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "Task ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TaskStatusResponse"
            }
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks-results/{task_id}": {
      "get": {
        "summary": "Get paginated results",
        "description": "Get paginated results for completed task",
        "tags": ["tasks"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "Task ID"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "type": "integer",
            "default": 100,
            "minimum": 1,
            "maximum": 100,
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated response",
            "schema": {
              "$ref": "#/definitions/PaginatedResponse"
            }
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks-with-webhook": {
      "post": {
        "summary": "Create verification task with webhook",
        "description": "Create task for email verification with callback webhook notification",
        "tags": ["tasks"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WebhookRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TaskIDResponse"
            }
          },
          "400": {
            "description": "Invalid request format or parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/swagger/": {
      "get": {
        "summary": "Get Swagger UI",
        "tags": ["documentation"],
        "produces": ["text/html"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Prometheus Metrics",
        "description": "Expose application metrics in Prometheus format",
        "tags": ["monitoring"],
        "produces": ["text/plain"],
        "responses": {
          "200": {
            "description": "Prometheus metrics",
            "schema": {
              "type": "string",
              "example": "# HELP http_requests_total Total HTTP requests\n# TYPE http_requests_total counter\nhttp_requests_total{method=\"GET\",path=\"/tasks\",status=\"200\"} 42"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "FullAPIKeyDetails": {
      "type": "object",
      "properties": {
        "api_key": {"type": "string"},
        "type": {"type": "string"},
        "remaining": {"type": "integer"},
        "used": {"type": "integer"},
        "initial": {"type": "integer"},
        "created_at": {"type": "string"},
        "expires_at": {"type": "string"},
        "last_topup": {"type": "string"}
      }
    },
    "KeyUpdateRequest": {
      "type": "object",
      "properties": {
        "add_checks": {"type": "integer"},
        "extend_days": {"type": "integer"}
      }
    },
    "CreateKeyRequest": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "enum": ["pay_as_you_go", "monthly"],
          "example": "monthly"
        },
        "initial_checks": {
          "type": "integer",
          "minimum": 1,
          "example": 1000
        },
        "validity_days": {
          "type": "integer",
          "minimum": 1,
          "example": 30
        }
      },
      "required": ["type", "initial_checks"]
    },
    "UpdateKeyRequest": {
      "type": "object",
      "properties": {
        "add_checks": {
          "type": "integer",
          "minimum": 1
        },
        "extend_days": {
          "type": "integer",
          "minimum": 1
        }
      }
    },
    "APIKeyResponse": {
      "type": "object",
      "properties": {
        "api_key": {
          "type": "string",
          "example": "abc123def456"
        },
        "type": {
          "type": "string",
          "example": "monthly"
        },
        "remaining_checks": {
          "type": "integer",
          "example": 950
        },
        "total_checks": {
          "type": "integer",
          "example": 1000
        },
        "expires_at": {
          "type": "string",
          "format": "date-time",
          "example": "2024-12-31T23:59:59Z"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "example": "2024-01-01T00:00:00Z"
        }
      }
    },
    "CacheStatusResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "integer",
          "example": 1425,
          "description": "Total cached items"
        },
        "memory": {
          "type": "string",
          "example": "12.34 MB",
          "description": "Memory usage"
        },
        "hits": {
          "type": "integer",
          "example": 24500,
          "description": "Cache hit count"
        },
        "misses": {
          "type": "integer",
          "example": 1780,
          "description": "Cache miss count"
        }
      }
    },
    "Request": {
      "type": "object",
      "properties": {
        "emails": {
          "type": "array",
          "items": {
            "type": "string",
            "maxLength": 254
          },
          "maxItems": 10000,
          "description": "Array of email addresses (maximum 10,000). Each email address must not exceed 254 characters."
        }
      },
      "description": "Request object containing a list of email addresses to verify."
    },
    "TaskIDResponse": {
      "type": "object",
      "properties": {
        "task_id": {
          "type": "string",
          "example": "123456789"
        }
      }
    },
    "TaskStatusResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "completed"
        },
        "total_results": {
          "type": "integer",
          "example": 1500
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "example": "2024-01-15T14:30:00Z"
        },
        "total_pages": {
          "type": "integer",
          "example": 15
        }
      }
    },
    "PaginatedResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EmailReport"
          }
        },
        "page": {
          "type": "integer",
          "example": 2
        },
        "per_page": {
          "type": "integer",
          "example": 100
        },
        "total": {
          "type": "integer",
          "example": 1500
        }
      }
    },
    "EmailReport": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "test@example.com"
        },
        "valid": {
          "type": "boolean",
          "example": true
        },
        "disposable": {
          "type": "boolean",
          "example": false
        },
        "exists": {
          "type": "boolean",
          "example": true
        },
        "mx": {
          "$ref": "#/definitions/MXStats"
        },
        "permanent_error": {
          "type": "boolean",
          "example": false
        },
        "error_category": {
          "type": "string",
          "example": "server_error"
        },
        "ttl": {
          "type": "integer",
          "example": 3600
        },
        "smtp_error": {
          "type": "string",
          "example": "550 Mailbox not found"
        }
      }
    },
    "MXStats": {
      "type": "object",
      "properties": {
        "valid": {
          "type": "boolean",
          "example": true
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MXRecord"
          }
        },
        "error": {
          "type": "string",
          "example": "MX lookup failed"
        }
      }
    },
    "MXRecord": {
      "type": "object",
      "properties": {
        "host": {
          "type": "string",
          "example": "mail.example.com"
        },
        "priority": {
          "type": "integer",
          "example": 10
        },
        "ttl": {
          "type": "integer",
          "example": 3600
        }
      }
    },
    "WebhookRequest": {
      "type": "object",
      "properties": {
        "emails": {
          "$ref": "#/definitions/Request/properties/emails"
        },
        "webhook": {
          "$ref": "#/definitions/WebhookConfig"
        }
      },
      "required": ["emails", "webhook"]
    },
    "WebhookConfig": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string",
          "format": "uri",
          "example": "https://api.example.com/webhook",
          "description": "Webhook URL for status notifications"
        },
        "ttl": {
          "type": "string",
          "pattern": "^\\d+[smh]$",
          "example": "1h",
          "description": "Webhook TTL (e.g., 30s, 5m, 1h)"
        },
        "retries": {
          "type": "integer",
          "minimum": 1,
          "maximum": 5,
          "example": 3,
          "description": "Max retry attempts for failed notifications"
        },
        "secret": {
          "type": "string",
          "example": "my-secret-key",
          "description": "HMAC signature secret (optional)"
        }
      },
      "required": ["url", "ttl", "retries"]
    },
    "WebhookResponse": {
      "type": "object",
      "properties": {
        "task_id": {
          "type": "string",
          "example": "1745264294247024000"
        },
        "status": {
          "type": "string",
          "example": "completed"
        },
        "results": {
          "type": "integer",
          "example": 1
        },
        "attempts": {
          "type": "integer",
          "example": 0,
          "description": "Number of delivery attempts made"
        },
        "ttl": {
          "type": "string",
          "example": "1h",
          "description": "Original webhook TTL"
        },
        "lifetime": {
          "type": "string",
          "example": "2.299709ms",
          "description": "Total task processing time"
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}